Buscar paquete ahora busca por ciudades
